Limit Hypothesis to <6.38
Otherwise in the tests we might get:

FloatingPointError: Got allow_subnormal=True, but we can't represent subnormal
floats right now, in violation of the IEEE-754 floating-point specification.
This is usually because something was compiled with -ffast-math or a similar
option, which sets global processor state.

Since this is because some library (lxml, gevent, psycopg2, you name it) is
compiled with -ffast-math; but since those libraries are been with us for a
long, long time, and our floating operations will rarely imply too big or too
small numbers; I think we can live keep with our lives having subnormal floats
in tests.

I'm also limiting hypothesis generated floats to a smaller range and 32 bits
to reduce the likelihood of this error in hypothesis driven tests.